Add simple validation for BuildConfig
BuildConfig must not be empty
